fix(ios): ScopedVaultPath deinit backstop + accurate idempotence doc
Review fixups: deinit releases a forgotten scope (exactly-once preserved);
doc states the @MainActor serial-call assumption instead of implying intrinsic
race-safety; move VaultSelectionError equality test to its own file (one test
file per type).
